Rewrote exec.asm, added pong.asm.
Total overhaul of exec.asm to use this fancy new technology called arrays, and to rely totally on jumps rather than a mixture of calls and jumps that damaged the stack. Changed word variables to byte variables where possible, saved 34 bytes of ram with this which then got immediately used us by rewriting some sentences to make more sense. Scene2 is now totally useless, but not adandoned so I'll keep it included. The game no longer hangs indefinitely when failing to save or loading the game. Added pong.asm, my personal take on pong (yes it is as I intended it, it isn't just a neatly arranged pile of salvage). Reworked files as necessary to facilitate the new exec.asm, nothing major in most cases. The Uinput routine needed fairly significant changing but otherwise it was just switching a call to a jump, adding the new return variable and flow control.