Added PS/2 mouse drivers, rewrote paint
Finally managed to write a functioning PS/2 mouse driver, and then rewrote paint to support the mouse. For now keyboard movement has been suspended but I doubt anyone will miss it too badly. Still needs a lot of work but I was excited to push the new drivers, so it's more of a proof of concept of what could be for now. The next big steps are a proper graphical cursor and changing colour at will. I plan to make z/x pen down and pen up as before too. On real hardware I had issues with the keyboard not always being usable after quitting and also the entirety of my monitor doesn't appear addressable, but that seems unavoidable since it's like 1024x768 and I'm in a 640x350 video mode. Also, the bug with turning off is fixed.
I forgot to write commit changes for last time, I added os.bin which is the compiled version of what I'm uploading and also added comments and standardised formatting throughout. Also removed unecessary pusha and popa from the piano.