Optimised print.asm and game.asm a lot
I reordered the game stats so now they can be dealt with much more efficiently, the battle stats are now ordered so they can be looked up and printed very easily and efficiently, and the other game stats are now in Binary Coded Decimal™ to make use of the relevant instructions. 452 bytes have been saved overall, and things are a whole lot cleaner now. There is still work to be done, I intend to rework how colours and bulk spaces are dealt with for printing, and I need to go through the second half of game.asm. I also need to test these new changes on real hardware.