Rewrote wait.asm, removed more pusha/popa, fixed saving, added debug tool.
I totally rewrote wait.asm so now pausing relies on the Programmable Interrupt Timer (PIT) instead of using a BIOS interrupt. The new  method is much more accurate I've found, as the old interrupt method would have inconsistencies of up to 4% from the expected measurements, but more importantly it's more widely supported and further reduces reliance on the BIOS. I made sure that the old wait methods could be exactly replicated, mostly so the beeping stays the same, but it's probably useful to have the waits happen in human units rather than arbitrary clock ticks. I removed some more pusha and popa instructions where I found them, as I'm still aiming for 8086 support (I would like to try testing on an IBM PC or an emulator, but I'm unsure where to get access to one or how accurate the emulation is), and I also fixed the bug where saving wasn't working, which was due to me writing 0x20 when I wanted 20d, or 0x14, to load in 20 sectors from the boot device. I also added a debugging tool that displays the contents of the 4 general purpose registers on the screen (in a human readable format) and then hangs the system. This update adds 236 bytes in total.b