Rewrote pong (again), commented out unused modules
I realised that space was starting to get limited, so I've uncommented load.asm, Scene2.asm and fight.asm since there's no way to actually access them currently. I totally rewrote pong since my old pong code was very poorly written, and decided to rewrite it using no conditional branching (except the necessary one to return to the menu) in order to get used to avoiding branching, as this is often the cause of my poor code. I'm not saying it's now perfect, in fact there are points where a conditional branch would almost certainly be a cleaner solution, but it's definitely more compact than it was, and is also now in 320x200 graphics mode. I've also managed to avoid using multiplication, and have used self-modifying code, so I'm happy with the range of techniques I've managed to use. 1467 bytes have been saved in this update, although not including the commenting out this is 390 bytes.